+++ tests/test_models.py
@@ -1,4 +1,5 @@
 import os
+import uuid
 
 from scim2_models import BulkRequest
 from scim2_models import BulkResponse
@@ -133,3 +134,32 @@ def test_everything_is_optional():
     ]
     for model in models:
         model()
+
+
+def test_case_sensitivity():
+    """RFC7643 §2.1 indicates that attribute names should be case insensitive.
+
+    Attribute names are case insensitive and are often "camel-cased"
+    (e.g., "camelCase").
+
+    Reported by issue #39.
+    """
+
+    payload = {
+        "UserName": "UserName123",
+        "Active": True,
+        "DisplayName": "BobIsAmazing",
+        "schemas": ["urn:ietf:params:scim:schemas:core:2.0:User"],
+        "externalId": uuid.uuid4().hex,
+        "name": {
+            "formatted": "Ryan Leenay",
+            "familyName": "Leenay",
+            "givenName": "Ryan",
+        },
+        "emails": [
+            {"Primary": True, "type": "work", "value": "testing@bob.com"},
+            {"Primary": False, "type": "home", "value": "testinghome@bob.com"},
+        ],
+    }
+    user = User.model_validate(payload)
+    assert user.display_name == "BobIsAmazing"